refactor: 数値プロパティをdoubleから32bit整数(int)に変更する
小数を必要とする属性は登場しない前提のため、propsの数値表現をintに揃えた。
PropertyValue.Number/PropertyDef(Range/Stage.Min)/ContributionDef.Amount、
WorldObjectのGetNumber/AddNumber/GetEffectiveValue、Viewsの各アクセサを
すべてintに変更。slot側のcapacity/weight_rateはpropsではなく端数を許容する
コンテナ設定値のためdoubleのまま残し、重さの伝播(Containment.PropagateWeight)は
weight_rateとの乗算結果を各階層へ加算する直前にだけ丸める。

WorldCodex.schema.jsonのprop_value/range_value/stage min/modify・accumulate・add
の値もnumberからintegerに厳格化し、小数値が拒否されることをjsonschemaで確認した。
GameElementDefinition.mdにも数値プロパティが32bit整数のみである旨を明記し、
6.4節のサンプル(temperature: 0.1/0.2)を整数値に更新した。
ContributionTests.csの体温上昇サンプルも整数値(36→37)に合わせて更新し、
dotnet testで既存14件+Views4件、計18件全てパスすることを確認した。

+++ Assets/Scripts/WorldCodex/Defs/PropertyDef.cs
@@ -6,10 +6,10 @@ namespace UnmappedIsland.Codex.Defs
 {
     public readonly struct PropertyRange
     {
-        public readonly double Min;
-        public readonly double Max;
+        public readonly int Min;
+        public readonly int Max;
 
-        public PropertyRange(double min, double max)
+        public PropertyRange(int min, int max)
         {
             Min = min;
             Max = max;
@@ -45,9 +45,9 @@ namespace UnmappedIsland.Codex.Defs
         public string Name { get; }
 
         /// <summary>下限。null は最下段（それより下の残り全ての値を拾う、6.4節）。</summary>
-        public double? Min { get; }
+        public int? Min { get; }
 
-        public PropertyStage(string name, double? min)
+        public PropertyStage(string name, int? min)
         {
             Name = name;
             Min = min;
@@ -112,7 +112,7 @@ namespace UnmappedIsland.Codex.Defs
         /// 現在値が該当する段階を返す。min:null の段階は「他のどの段階にも該当しない場合」のフォールバックであり、
         /// リスト中の位置には依存しない（11.2節のサンプルでは broken(min:null) が intact(min:1) より後に書かれている）。
         /// </summary>
-        public PropertyStage ResolveStage(double currentValue)
+        public PropertyStage ResolveStage(int currentValue)
         {
             PropertyStage fallback = null;
             PropertyStage best = null;
